Organized Code, added success screen, added auth
