+++ smactest-app/src/pages/Questions.css
@@ -1,9 +1,16 @@
+.questions-container {
+    max-width:800px;
+    margin: 20px auto 20px auto;
+}
+
 input {
     padding:10px;
     border-radius:10px;
     color: #b3002d;
     border:0.5px solid #b3002d;
     border-style:double;
+    width:100%;
+    font-weight: bold;
 }
 
 input:focus {
