+++ smactest-app/src/pages/rankData/rankChoices.js
@@ -1,0 +1,65 @@
+const rankChoices = {
+    purple: [
+      [
+        'A sophisticated game of Tag',
+        'An all-out fight with rules',
+        'A chance to hurt someone using Kung-fu',
+        'A test of one\'s might as a fighter'
+      ],
+      [
+        'A chance to hurt someone using Karate',
+        'Applications based on real-life scenarios',
+        'An exercise in skilled paranoia',
+        'A different way to combine basics'
+      ],
+      [
+        'The Heel',
+        'The Side Blade of the foot',
+        'The \'Ball\' of the foot',
+        'The Top or \'Bridge\' of the foot'
+      ],
+      [
+        'Front, Crescent, Spear',
+        'Round, Rev. Crescent, Hook',
+        'Side to Front, Round, Hook',
+        'Stepping stool, X-Step, Hook'
+      ],
+      [
+        '1 - 2',
+        '1 - 4',
+        '1 - 6',
+        '1 - 8'
+      ],
+      [
+        'Chin-Na & Shuai Jiao',
+        'Judo & Ninjutsu',
+        'Brazilian Jujutsu & MMA',
+        'All of the above'
+      ],
+      [
+        '7',
+        '6',
+        '5',
+        '4'
+      ],
+      [
+        'An Indian dish of flat bread & rice',
+        'A Buddhist monk/warrior from India',
+        'The founder of the Shaolin Temple',
+        'I don\'t know'
+      ],
+      [
+        'Okinawa',
+        'Japan',
+        'China',
+        'Tibet'
+      ],
+      [
+        'The back of the hand',
+        'The bottom two finger knuckles',
+        'Tips of first two fingers',
+        'The bottom of the hand'
+      ]
+    ]
+}
+export default rankChoices
