+++ smactest-app/src/pages/rankData/rankPrices.js
@@ -1,0 +1,4 @@
+const rankPrices = {
+    Purple: 60
+}
+export default rankPrices
