+++ smactest-app/src/pages/rankData/rankQuestions.js
@@ -1,0 +1,15 @@
+const rankQuestions = {
+    purple: [
+      'What is Sparring?', 
+      'What is Self-Defense?', 
+      'What part of the foot is used for a Round Kick?', 
+      'What are the 3 kicks I learned in Purple Belt?', 
+      'How many of the 8pt Open Hand Blocks have you learned?', 
+      'The primary arts used in our grappling are:', 
+      'How many Animals are part of the Shaolin Training system?',
+      'Bodhidharma (or Tamo) is:', 
+      'The Shaolin Temple is located in:', 
+      'The Striking point of Immortal Hand is:'
+    ]
+}
+export default rankQuestions
